requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1
Werkzeug==2.3.7
python-dotenv==1.0.0
alembic==1.13.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9

# PDF processing libraries
PyPDF2==3.0.1
pdfplumber==0.10.3
tabula-py==2.9.0

# Additional text processing
chardet==5.2.0

# Production server and caching
gunicorn==21.2.0
Flask-Caching==2.1.0

# Security and monitoring
Flask-Talisman==1.1.0
